# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Qwen-2.5-VL API",
        "description": "API for Qwen-2.5-VL model with image and text inputs",
        "version": "0.0.1"
    },
    "paths": {
...
        "/cases/": {
            "get": {
                "summary": "Get Cases",
                "operationId": "get_cases_cases__get",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Case",
                "operationId": "create_case_cases__post",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "case_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Case ID for the case",
                            "title": "Case Id"
                        },
                        "description": "Case ID for the case"
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Patient ID for the case",
                            "title": "Patient Id"
                        },
                        "description": "Patient ID for the case"
                    },
                    {
                        "name": "case_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Name of the case",
                            "title": "Case Name"
                        },
                        "description": "Name of the case"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Description of the case",
                            "title": "Description"
                        },
                        "description": "Description of the case"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "description": "Tags for the case",
                            "default": [],
                            "title": "Tags"
                        },
                        "description": "Tags for the case"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cases/{case_id}": {
            "get": {
                "summary": "Get Case",
                "operationId": "get_case_cases__case_id__get",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "case_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Case Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Case",
                "operationId": "update_case_cases__case_id__put",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "case_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Case Id"
                        }
                    },
                    {
                        "name": "case_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Updated name of the case",
                            "title": "Case Name"
                        },
                        "description": "Updated name of the case"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Updated description of the case",
                            "title": "Description"
                        },
                        "description": "Updated description of the case"
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "description": "Updated tags for the case",
                            "title": "Tags"
                        },
                        "description": "Updated tags for the case"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Case",
                "operationId": "delete_case_cases__case_id__delete",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "case_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Case Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/history/{session_id}": {
            "get": {
                "summary": "Get History",
                "operationId": "get_history_history__session_id__get",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete History",
                "operationId": "delete_history_history__session_id__delete",
                "security": [
                    {
                        "JWTBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "JWTBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}